pyproject.toml
[project]
name = "smart-cane"
version = "0.1.0"
description = ""
authors = [
    { name = "yoni_ash", email = "yonatha12345678910@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "sounddevice (>=0.5.1,<0.6.0)",
    "numpy",
    "openai-whisper (==20240930)",
    "pyttsx3 (>=2.98,<3.0)",
    "ultralytics (>=8.3.73,<9.0.0)",
    "noisereduce (>=3.0.3,<4.0.0)",
    "vosk (>=0.3.45,<0.4.0)",
    "pygame (>=2.6.1,<3.0.0)",
    "gtts (>=2.5.4,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
notebook = "^7.3.2"

